feat: split up multi hop ui components into files (#5575)
* feat: split up multi hop ui components into files

* chore: actioned code review feedback
