+++ src/components/MultiHopTrade/components/MultiHopTradeConfirm/types.ts
@@ -1,0 +1,4 @@
+export enum TradeType {
+  Bridge = 'Bridge',
+  Swap = 'Swap',
+}
